Optimize exports, add unit test for exports
