+++ src/support/defs.ts
@@ -1,0 +1,11 @@
+export type Progress = {
+  x: number;
+  y: number;
+};
+
+export type Options = {
+  /** Mirror the vertical scroll position */
+  vertical: boolean;
+  /** Mirror the horizontal scroll position */
+  horizontal: boolean;
+};
