--- src/support/devs.ts
@@ -1,4 +1,0 @@
-export type Progress = {
-  x: number;
-  y: number;
-};
